Menerapkan authorization

+++ src/api/notes/handler.js
@@ -15,9 +15,13 @@ class NotesHandler {
   async postNoteHandler(request, h) {
     try {
       this._validator.validateNotePayload(request.payload);
+
       const { title = 'untitled', body, tags } = request.payload;
-    
-      const noteId = await this._service.addNote({ title, body, tags });
+      const { id: credentialId } = request.auth.credentials;
+
+      const noteId = await this._service.addNote({
+        title, body, tags, owner: credentialId,
+      });
     
       const response = h.response({
         status: 'success',
@@ -49,8 +53,9 @@ class NotesHandler {
     }
   }
 
-  async getNotesHandler() {
-    const notes = await this._service.getNotes();
+  async getNotesHandler(request) {
+    const { id: credentialId } = request.auth.credentials;
+    const notes = await this._service.getNotes(credentialId);
     return {
       status: 'success',
       data: {
@@ -62,6 +67,9 @@ class NotesHandler {
   async getNoteByIdHandler(request, h) {
     try {
       const { id } = request.params;
+      const { id: credentialId } = request.auth.credentials;
+
+      await this._service.verifyNoteOwner(id, credentialId);
       const note = await this._service.getNoteById(id);
       return {
         status: 'success',
@@ -90,11 +98,13 @@ class NotesHandler {
     }
   }
 
-  putNoteByIdHandler(request, h) {
+  async putNoteByIdHandler(request, h) {
     try {
       this._validator.validateNotePayload(request.payload);
       const { id } = request.params;
+      const { id: credentialId} = request.auth.credentials;
     
+      await this._service.verifyNoteOwner(id, credentialId);
       this._service.editNoteById(id, request.payload);
     
       return {
@@ -122,9 +132,12 @@ class NotesHandler {
     }
   }
 
-  deleteNoteByIdHandler(request, h) {
+  async deleteNoteByIdHandler(request, h) {
     try {
       const { id } = request.params;
+      const { id: credentialId } = request.auth.credentials;
+
+      await this._service.verifyNoteOwner(id, credentialId);
       this._service.deleteNoteById(id);
       return {
         status: 'success',
